requirements.txt
terminaltables=3.1.0
python-dotenv==0.10.1
requests==2.20.0
